downgrade frb to 1.82.6

+++ lib/bloc/redeem_onchain_funds/redeem_onchain_funds_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:breez_sdk/breez_sdk.dart';
-import 'package:breez_sdk/sdk.dart';
+import 'package:breez_sdk/bridge_generated.dart';
 import 'package:breez_translations/breez_translations_locales.dart';
 import 'package:c_breez/bloc/redeem_onchain_funds/redeem_onchain_funds_state.dart';
 import 'package:c_breez/models/models.dart';
@@ -61,7 +61,7 @@ class RedeemOnchainFundsBloc extends Cubit<RedeemOnchainFundsState> {
     ];
     final feeOptions = await Future.wait(
       List.generate(3, (index) async {
-        final recommendedFee = recommendedFeeList.elementAt(index).toInt();
+        final recommendedFee = recommendedFeeList.elementAt(index);
         final req = PrepareRedeemOnchainFundsRequest(toAddress: toAddress, satPerVbyte: recommendedFee);
         final txFeeSat = await prepareRedeemOnchainFunds(req);
 
@@ -82,7 +82,7 @@ class RedeemOnchainFundsBloc extends Cubit<RedeemOnchainFundsState> {
     try {
       final resp = await _breezSDK.prepareRedeemOnchainFunds(req: req);
       _log.info("Redeem onchain funds txFee: ${resp.txFeeSat}, with tx weight ${resp.txWeight}");
-      return resp.txFeeSat.toInt();
+      return resp.txFeeSat;
     } catch (e) {
       _log.severe("Failed to redeem onchain funds", e);
       rethrow;
